+++ R/methane.R
@@ -53,7 +53,7 @@ connected <- .isConnected(file_url)
 if (!connected) {message("Retrieving remote data requires connectivity to source."); return(invisible(NULL))}
 
 dl <- tempfile()
-status <- tryCatch({  download.file(file_url, dl) }, error = function(e) {TRUE}, error = function(e) {TRUE} )
+status <- tryCatch(suppressWarnings(download.file(file_url, dl)), error = function(e) 1L)
 if (status!=0L) {message("Unable to access remote resource."); return(invisible(NULL))}
 
 ch4 <- suppressMessages( read_table(dl, col_names = FALSE, skip = 65) )